Fixed Turret Math Conversion
So basically, our turret aiming system already knows how to aim toward a point as long as it has an absolute position and heading. So instead of doing our own math to find those values, we can just use the SDK’s built in AprilTag detection, which already gives a pose estimate. We just treat that pose as the turret’s position rather than the robot’s. That way, the vision system and the odometry system both use the same aiming logic: the only difference is that odometry calculates turret position from the robot, while vision gets it directly from the SDK

Ibn Lehoofy